+++ v4/localenv.go
@@ -318,6 +318,13 @@ func (e *LocalEnv) ReplacePositionalParams(values ...string) int {
 	return replacePositionalParams(e, values...)
 }
 
+// ResetPositionalParams deletes $1, $2 etc etc from the environment.
+//
+// It also sets the special parameter $# to 0.
+func (e *LocalEnv) ResetPositionalParams() {
+	resetPositionalParams(e)
+}
+
 // SetPositionalParams sets $1, $2 etc etc to the given values.
 //
 // Any existing positional parameters are overwritten, up to len(values).